Move Span dataclasses to MLflow entities
