+++ mlflow/entities/__init__.py
@@ -20,6 +20,12 @@ from mlflow.entities.run_inputs import RunInputs
 from mlflow.entities.run_status import RunStatus
 from mlflow.entities.run_tag import RunTag
 from mlflow.entities.source_type import SourceType
+from mlflow.entities.span import Span, SpanType
+from mlflow.entities.span_context import SpanContext
+from mlflow.entities.span_event import SpanEvent
+from mlflow.entities.span_status import SpanStatus
+from mlflow.entities.trace import Trace
+from mlflow.entities.trace_data import TraceData
 from mlflow.entities.trace_info import TraceInfo
 from mlflow.entities.trace_status import TraceStatus
 from mlflow.entities.view_type import ViewType
@@ -42,6 +48,13 @@ __all__ = [
     "InputTag",
     "DatasetInput",
     "RunInputs",
+    "Span",
+    "SpanContext",
+    "SpanEvent",
+    "SpanStatus",
+    "SpanType",
+    "Trace",
+    "TraceData",
     "TraceInfo",
     "TraceStatus",
     "_DatasetSummary",
