+++ tests/tracing/clients/test_local.py
@@ -1,6 +1,5 @@
-from mlflow.entities.trace_info import TraceInfo
+from mlflow.entities import SpanStatus, Trace, TraceInfo
 from mlflow.tracing.clients import InMemoryTraceClient
-from mlflow.tracing.types.model import Status, StatusCode, Trace
 
 
 def test_log_and_get_trace(monkeypatch):
@@ -13,7 +12,7 @@ def test_log_and_get_trace(monkeypatch):
                 experiment_id="test",
                 start_time=0,
                 end_time=1,
-                status=Status(StatusCode.OK),
+                status=SpanStatus(SpanStatus.StatusCode.OK),
                 attributes=[],
                 tags=[],
             ),
